changed netcons and synapses; removed complex synapses
